requirements.txt
streamlit==1.35.0
requests==2.32.5
plotly==5.22.0
pandas==2.2.2
fpdf2==2.7.9
python-nmap==0.7.1
